pyproject.toml
[project]
name = "mcp-file-search"
version = "1.0.0"
description = "MCP server for intelligent file search with natural language processing"
authors = [
    { name = "Lucy", email = "lucy@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
